Create a PizzaNinja model
