+++ app/Http/Controllers/CollectionController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Inscription;
+use App\Models\PizzaNinja;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\View\View;
 
@@ -10,7 +10,7 @@ class CollectionController extends Controller
 {
     public function __invoke(): View
     {
-        $inscriptions = Inscription::whereHas('collection', function (Builder $query) {
+        $inscriptions = PizzaNinja::whereHas('collection', function (Builder $query) {
             $query->where('slug', 'pizza-ninjas');
         })
             ->orderBy('name')
