+++ app/Http/Controllers/InscriptionController.php
@@ -2,12 +2,12 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Inscription;
+use App\Models\PizzaNinja;
 use Illuminate\View\View;
 
 class InscriptionController extends Controller
 {
-    public function __invoke(Inscription $inscription): View
+    public function __invoke(PizzaNinja $inscription): View
     {
         if (! $inscription->collection || $inscription->collection->slug !== 'pizza-ninjas') {
             abort(404);
