+++ app/Http/Controllers/PizzaNinjaController.php
@@ -2,14 +2,14 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Inscription;
+use App\Models\PizzaNinja;
 use Illuminate\Http\RedirectResponse;
 
 class PizzaNinjaController extends Controller
 {
     public function __invoke(int $id): RedirectResponse
     {
-        $inscription = Inscription::where('name', 'LIKE', "%#{$id}")
+        $inscription = PizzaNinja::where('name', 'LIKE', "%#{$id}")
             ->first();
 
         if (! $inscription) {
