+++ app/Http/Controllers/SearchController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Inscription;
+use App\Models\PizzaNinja;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
@@ -22,7 +22,7 @@ class SearchController extends Controller
 
         $searchQuery = Str::of($validated['query'])->after('#')->toString();
 
-        $inscription = Inscription::whereRelation('collection', 'slug', 'pizza-ninjas')
+        $inscription = PizzaNinja::whereRelation('collection', 'slug', 'pizza-ninjas')
             ->where(function (Builder $query) use ($searchQuery) {
                 $query->where('name', 'LIKE', "%#{$searchQuery}")
                     ->orWhere('inscription_id', $searchQuery);
