+++ app/View/Components/RandomNinjas.php
@@ -2,10 +2,12 @@
 
 namespace App\View\Components;
 
+use App\Models\PizzaNinja;
 use App\Models\Inscription;
-use Illuminate\Contracts\View\View;
-use Illuminate\Database\Eloquent\Collection;
 use Illuminate\View\Component;
+use Illuminate\Contracts\View\View;
+use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\Collection;
 
 class RandomNinjas extends Component
 {
@@ -13,7 +15,12 @@ class RandomNinjas extends Component
 
     public function __construct()
     {
-        $this->records = Inscription::inRandomOrder()->limit(5)->get();
+        $this->records = PizzaNinja::whereHas('collection', function (Builder $query) {
+            $query->where('slug', 'pizza-ninjas');
+        })
+            ->inRandomOrder()
+            ->limit(5)
+            ->get();
     }
 
     public function render(): View
